apps/worker/pyproject.toml
[project]
name = "yt-growth-worker"
version = "0.2.0"
description = "Trending niches discovery worker - ingestion, snapshotting, embedding, clustering, and scoring pipeline"
requires-python = ">=3.11"
dependencies = [
    "psycopg[binary]>=3.1.0",
    "openai>=1.0.0",
    "numpy>=1.24.0",
    "scikit-learn>=1.3.0",
    "hdbscan>=0.8.33",
    "umap-learn>=0.5.4",
    "click>=8.1.0",
    "python-dotenv>=1.0.0",
    "tenacity>=8.2.0",
    "pgvector>=0.2.0",
    "requests>=2.31.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/worker"]

[project.scripts]
worker = "worker.__main__:cli"
